only rm if dir exists

+++ resetProject.sh
@@ -4,7 +4,9 @@ curDir=`pwd`
 projDir=$1
 cd $projDir
 git reset --hard
-rm -r node_modules
+if [ -d node_modules ]; then
+	rm -r node_modules
+fi
 npm install
 npm run build --if-present
 cd $curDir
